Backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-restframework==0.0.1
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
